+++ src/controllers/api/v1/tickets.js
@@ -1982,7 +1982,7 @@ apiTickets.subscribe = function (req, res) {
  * @apiDescription Gets all ticket tags
  * @apiVersion 0.1.6
  * @apiGroup Ticket
- * @apiHeader {string} accesstoken The access token for the logged in user
+ * @apiHeader {string} accesstoken The access token for the logged-in user
  *
  * @apiExample Example usage:
  * curl -H "accesstoken: {accesstoken}" -l http://localhost/api/v1/tickets/tags
@@ -2059,7 +2059,7 @@ apiTickets.getOverdue = function (req, res) {
       ],
       function (err, overdueTickets) {
         if (err) return res.status(400).json({ success: false, error: err.message })
-
+        if (!overdueTickets) overdueTickets = []
         return res.json({ success: true, tickets: overdueTickets })
       }
     )