+++ src/controllers/api/v2/accounts.js
@@ -62,7 +62,7 @@ accountsApi.create = async function (req, res) {
   const chance = new Chance()
 
   try {
-    if (postData.password == undefined || postData.passwordConfirm == undefined) throw new Error('Password length is too short.')
+    if (!postData.password || !postData.passwordConfirm) throw new Error('Password length is too short.')
 
     // SETTINGS
     const SettingsUtil = require('../../../settings/settingsUtil')