+++ src/mailer/mailCheck.js
@@ -139,12 +139,6 @@ function bindImapError() {
 
 function bindImapReady() {
   try {
-    mailCheck.Imap.on('end', function () {
-      handleMessages(mailCheck.messages, function () {
-        mailCheck.Imap.destroy()
-      })
-    })
-
     mailCheck.Imap.on('ready', function () {
       openInbox(function (err) {
         if (err) {
@@ -177,6 +171,7 @@ function bindImapReady() {
 
                 f.on('message', function (msg) {
                   msg.on('body', function (stream) {
+                    var message = {}
                     var buffer = ''
                     stream.on('data', function (chunk) {
                       buffer += chunk.toString('utf8')
@@ -244,20 +239,14 @@ function bindImapReady() {
                 })
 
                 f.on('end', function () {
-                  async.series(
-                    [
-                      function (cb) {
-                        mailCheck.Imap.addFlags(results, flag, cb)
-                      },
-                      function (cb) {
-                        mailCheck.Imap.closeBox(true, cb)
-                      }
-                    ],
-                    function (err) {
-                      if (err) winston.warn(err)
-                      return next()
-                    }
-                  )
+                  mailCheck.Imap.addFlags(results, flag, function() {
+                    mailCheck.Imap.closeBox(true, function () {
+                      mailCheck.Imap.end();
+                      handleMessages(mailCheck.messages, function () {
+                        mailCheck.Imap.destroy()
+                      })
+                    })
+                  })
                 })
               }
             ],